move tests/deepsparse/loggers to tests/deepsparse/legacy/loggers and pass tests
