+++ tests/deepsparse/legacy/loggers/metric_functions/computer_vision/test_built_ins.py
