+++ tests/deepsparse/legacy/loggers/metric_functions/helpers/test_config_generation.py
