+++ tests/deepsparse/legacy/loggers/metric_functions/predefined/predefined_logs/image_classification.txt
