+++ tests/deepsparse/legacy/loggers/metric_functions/predefined/predefined_logs/object_detection.txt
