+++ tests/deepsparse/legacy/loggers/metric_functions/predefined/predefined_logs/segmentation.txt
