+++ tests/deepsparse/legacy/loggers/metric_functions/predefined/predefined_logs/sentiment_analysis_1.txt
