+++ tests/deepsparse/legacy/loggers/metric_functions/predefined/predefined_logs/zero_shot_text_classification_1.txt
