+++ tests/deepsparse/legacy/loggers/metric_functions/test_built_ins.py
