+++ tests/deepsparse/legacy/loggers/test_function_logger.py
