+++ tests/deepsparse/legacy/loggers/test_python_logger.py
