+++ tests/deepsparse/legacy/loggers/test_str.py
